users API call is now paginated
change /allUsers API call to just /users due to the addition of it now being paginated. went to frontend AlumniCardGroup.js and changed the /allUsers call to just /users so the homepage is calling the right API. Ability to change the homepage to show a fixed size results that is returned from the /users json, meaning we can limit homepage to 10 users, for example.