API refactored code & cleanup & filtered API call created
Made a folder called API where there now exists a file for queries, routes, and controllers that are all utilized by the API. Server.js file now is short and sweet without the mess of the entire API declared within it. Added API call for filtering. It is not extensively tested