API router changed. Fetch statements added.
API user/:findId is now a POST method called /user which posts a body that contains a users email, and it returns the data from the db of that user. Fetch statement added for using this new API call so that when a user logs in, their data can be saved somewhere. Also started the fetch statement for creating a new user on form submission.